# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documents",
        "description": "API Documents description for Emotiv subscription",
        "contact": {
            "email": "dev@emotiv.com"
        },
        "license": {
            "name": "Emotiv@2018",
            "url": "https://www.emotiv.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Server dev"
        },
        {
            "url": "https://dev-services.emotiv.com/api",
            "description": "Server Staging"
        },
        {
            "url": "https://services.emotiv.com/api",
            "description": "Server Live"
        }
    ],
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "Example"
                ],
                "summary": "Get list users",
                "description": "Returns list users",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "New user info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "format": "int32",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Tan"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "Hanoi"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Example"
                ],
                "summary": "Add new users",
                "description": "Add new users",
                "operationId": "addUsers",
                "requestBody": {
                    "description": "User object info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "default": ""
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New user info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "address": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Example"
                ],
                "summary": "Get user detail",
                "description": "Returns user info",
                "operationId": "getUsersDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "format": "int32",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Tan"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "Hanoi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Example"
                ],
                "summary": "Update users",
                "description": "update users",
                "operationId": "updateUsers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User object info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "default": ""
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New user info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "address": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Example"
                ],
                "summary": "Delete users",
                "description": "Delete users",
                "operationId": "deleteUsers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User_id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    }
}